# backend/contest_training_system/docs/api.yml
openapi: 3.0.0
info:
  title: 竞赛集训系统 API
  description: '根据提供的Markdown文档生成的完整版竞赛集训系统RESTful API'
  version: '1.0.0'
servers:
  - url: https://api.example.com
    description: '生产服务器'
...
paths:
...
  /training_plan/create:
    post:
      tags:
        - '6. 训练计划'
      summary: '6.1 新建训练计划'
      description: '教师新建训练计划'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - startTime
                - end_time
                - contestIds
                - studentIds
              properties:
                title:
                  type: string
                  example: "春季集训"
                description:
                  type: string
                  example: "2025年春季集训计划"
                startTime:
                  type: string
                  format: date-time
                  example: "2025-03-01 09:00:00"
                end_time:
                  type: string
                  format: date-time
                  example: "2025-05-01 18:00:00"
                contestIds:
                  type: array
                  items:
                    type: integer
                  example: [101, 102]
                studentIds:
                  type: array
                  items:
                    type: integer
                  example: [1001, 1002]
      responses:
        '200':
          description: '创建成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "创建成功"
                  data:
                    type: object
                    properties:
                      planId:
                        type: integer
                        example: 301
  /training_plan/update:
    put:
      tags:
        - '6. 训练计划'
      summary: '6.2 修改训练计划'
      description: '教师修改训练计划'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - planId
              properties:
                planId:
                  type: integer
                  example: 301
                title:
                  type: string
                  example: "春季集训（修改版）"
                description:
                  type: string
                  example: "2025年春季集训计划修改版"
                startTime:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                contestIds:
                  type: array
                  items:
                    type: integer
                  example: [101, 102, 103]
                studentIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: '修改成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "修改成功"
                  data:
                    type: object
                    properties:
                      planId:
                        type: integer
                        example: 301
                      title:
                        type: string
                        example: "春季集训（修改版）"
  /training_plan/{planId}:
    get:
      tags:
        - '6. 训练计划'
      summary: '6.5 获取训练计划详情'
      description: '获取训练计划详细信息'
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
          example: 301
      responses:
        '200':
          description: '获取成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      planId:
                        type: integer
                      title:
                        type: string
                      description:
                        type: string
                      state:
                        type: string
                      startTime:
                        type: string
                        format: date-time
                      end_time:
                        type: string
                        format: date-time
                      creatorId:
                        type: integer
                      contests:
                        type: array
                        items:
                          type: object
                          properties:
                            contestId:
                              type: integer
                            title:
                              type: string
                            sequence:
                              type: integer
                            addedTime:
                              type: string
                              format: date-time
                            description:
                              type: string
                      students:
                        type: array
                        items:
                          type: object
                          properties:
                            userId:
                              type: integer
                            nickname:
                              type: string
                            enrolledAt:
                              type: string
                              format: date-time
    delete:
      tags:
        - '6. 训练计划'
      summary: '6.3 删除训练计划'
      description: '删除训练计划'
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
          example: 301
      responses:
        '200':
          description: '删除成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "删除成功"
                  data:
                    type: object
                    properties:
                      planId:
                        type: integer
                        example: 301
                      action:
                        type: string
                        example: "deleted"
  /training_plan/list:
    get:
      tags:
        - '6. 训练计划'
      summary: '6.4 查询训练计划列表'
      description: '查询训练计划列表'
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: state
          in: query
          schema:
            type: string
            enum: [UPCOMING, ONGOING, FINISHED]
        - name: creatorId
          in: query
          schema:
            type: integer
        - name: keyword
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [id_asc, id_desc, title_asc, title_desc, start_time_asc, start_time_desc, created_asc, created_desc]
            default: 'id_asc'
      responses:
        '200':
          description: '查询成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 20
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
                      plans:
                        type: array
                        items:
                          type: object
                          properties:
                            planId:
                              type: integer
                            title:
                              type: string
                            state:
                              type: string
                            startTime:
                              type: string
                              format: date-time
                            end_time:
                              type: string
                              format: date-time
                            creatorId:
                              type: integer
                            studentCount:
                              type: integer
                            contestCount:
                              type: integer

  # 8. 文件管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '在需要鉴权的接口的请求头中添加 "Authorization: Bearer <YOUR_TOKEN>"'